Fixes issues with project type.   Adds Twitter Search

+++ src/com/fenchurchtech/messaging/common/Encoder.java
@@ -15,7 +15,7 @@ public class Encoder {
             }
             else
             {
-                result.append('%' +  Integer.toHexString(symbol).toUpperCase());
+                result.append('%' +  Integer.toHexString((int)symbol).toUpperCase());
             }
         }
